pyproject.toml
[tool.poetry]
name = "spotify-song-sorter"
version = "0.1.0"
description = ""
authors = ["Marc Soler <93591805+marc-soler@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.3"
scikit-learn = "^1.6.0"
spotipy = "^2.24.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
